requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
requests
websocket-client
Pillow